Sprites and Scene update
I added a few tiles and an idle Hidekage. Anyone is welcome to make
changes. Was just testing positioning. The grassy area Hide is floating
over is 5 tiles (blocks) wide. If we have a map that is 3000 tiles wide
(sensible) then it'll take forever doing it by hand. We need to figure
out the quickest way to add blocks. I'll start the search.